# assets/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "charted-server",
        "description": "🐻‍❄️📦 Free, open source, and reliable Helm Chart registry made in Rust",
        "termsOfService": "https://charts.noelware.org/legal/tos",
        "contact": {
            "name": "Noelware, LLC.",
            "url": "https://noelware.org",
            "email": "team@noelware.org"
        },
        "license": {
            "name": "Apache 2.0 License",
            "url": "https://apache.org/licenses/LICENSE-2.0",
            "identifier": "Apache-2.0"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "https://charts.noelware.org/api/v{version}",
            "description": "Production Server",
            "variables": {
                "version": {
                    "default": "1",
                    "description": "Revision of the HTTP specification",
                    "enum": ["1"]
                }
            }
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": ["Users"],
                "summary": "Entrypoint to the Users API.",
                "operationId": "users",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/EntrypointResponse"
                    }
                }
            },
            "post": {
                "tags": ["Users"],
                "summary": "Creates a new user.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Request body for creating a new user on this instance. The `password` field can be omitted if the session backend isn't the local one.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/UserResponse"
                    },
                    "403": {
                        "description": "Instance doesn't allow registrations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            },
                                            "description": "List of errors that happened. This can be represented as a stacktrace"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "description": "Whether if this request was a success or not (always false)",
                                            "default": false
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Session backend required the `password` field or `email` was not a valid, proper email address",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            },
                                            "description": "List of errors that happened. This can be represented as a stacktrace"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "description": "Whether if this request was a success or not (always false)",
                                            "default": false
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "If a user already has the `username` or `email` taken.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            },
                                            "description": "List of errors that happened. This can be represented as a stacktrace"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "description": "Whether if this request was a success or not (always false)",
                                            "default": false
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Internal Server Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            },
                                            "description": "List of errors that happened. This can be represented as a stacktrace"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "description": "Whether if this request was a success or not (always false)",
                                            "default": false
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{idOrName}": {
            "get": {
                "tags": ["Users"],
                "summary": "Retrieve a single user by their ID or name.",
                "operationId": "getUserByIdOrName",
                "parameters": [
                    {
                        "name": "idOrName",
                        "in": "path",
                        "description": "Queries this resource by the entity's ID or name.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/NameOrUlid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/UserResponse"
                    },
                    "400": {
                        "description": "Invalid ID or name specified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            },
                                            "description": "List of errors that happened. This can be represented as a stacktrace"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "description": "Whether if this request was a success or not (always false)",
                                            "default": false
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            },
                                            "description": "List of errors that happened. This can be represented as a stacktrace"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "description": "Whether if this request was a success or not (always false)",
                                            "default": false
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NameOrUlid": {
                "oneOf": [
                    {
                        "type": "string",
                        "format": "ulid"
                    },
                    {
                        "$ref": "#/components/schemas/Name"
                    }
                ],
                "description": "`NameOrUlid` is a \"union\" enum that can represent either:\n\n* [`Name`]\n* [`Ulid`]"
            },
...
        },
        "responses": {
...
            "EntrypointResponse": {
                "description": "Response datatype that returns a `Entrypoint` object",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "required": ["success"],
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Entrypoint"
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Error"
                                    },
                                    "description": "The error trace for the request that was processed by\nthe API server.\n\nThe `success` field will always be set to `false` when\nthe `errors` field is avaliable."
                                },
                                "success": {
                                    "type": "boolean",
                                    "description": "Was the request that was processed a success?"
                                }
                            }
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "ApiKey": {
                "type": "apiKey",
                "in": "header",
                "name": "ApiKey"
            },
            "Basic": {
                "type": "http",
                "scheme": "basic",
                "description": "Allows to use basic authentication (base64 of `username:password`). Do note that some instances might have this disabled and it isn't recommended to be used in production environments."
            },
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Signed token by the server to safely authenticate yourself"
            }
        }
    },
...
}